review actions

+++ src/icrar/leap-accelerate/model/Integration.h
@@ -56,7 +56,7 @@ namespace icrar
 
         union
         {
-            std::array<int, 4> parameters; // index, 0, channels, baselines
+            std::array<size_t, 4> parameters; // index, 0, channels, baselines
             struct
             {
                 size_t index;
